% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPurpleairApiHistoryV2.R
\name{getPurpleairApiHistoryV2}
\alias{getPurpleairApiHistoryV2}
\title{getPurpleairApiHistoryV2}
\usage{
getPurpleairApiHistoryV2(
  sensorIndex = NULL,
  apiReadKey = NULL,
  startDate = NULL,
  endDate = NULL,
  average = NULL,
  fields = NULL,
  printFlag = FALSE
)
}
\arguments{
\item{sensorIndex}{PurpleAir sensor’s index.}

\item{apiReadKey}{PurpleAir API read key with access to historical data.}

\item{startDate}{The beginning date in the format "YYYY-MM-DD".}

\item{endDate}{The end date in the format "YYYY-MM-DD".}

\item{average}{The desired average in minutes, one of the following:
"0" (real-time), "10", "30", "60", "360" (6 hour), "1440" (1 day).}

\item{fields}{The "Fields" parameter specifies which 'sensor data fields' to
include in the response.
See: https://community.purpleair.com/t/api-history-fields-descriptions/}

\item{printFlag}{Logical. Indicates whether to print progress messages.}
}
\value{
Dataframe of PurpleAir history data of one or multiple sensors.
}
\description{
R function to download historical data from PurpleAir sensors from API.
Adapted from the getPurpleairApiHistory function
Source: https://github.com/willianflores/getPurpleairApiHistory
}
