% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purpleAirAPI.R
\name{validate_api_key}
\alias{validate_api_key}
\title{Helper function to validate the API key}
\usage{
validate_api_key(apiReadKey)
}
\arguments{
\item{apiReadKey}{PurpleAir API read key.}
}
\description{
Helper function to validate the API key
}
\keyword{internal}
