% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPurpleairApiHistoryV2.R
\name{getPurpleairApiHistoryV2}
\alias{getPurpleairApiHistoryV2}
\title{getPurpleairApiHistoryV2}
\usage{
getPurpleairApiHistoryV2(
  sensorIndex = NULL,
  apiReadKey = NULL,
  startTimeStamp = NULL,
  endTimeStamp = NULL,
  average = NULL,
  fields = NULL
)
}
\arguments{
\item{sensorIndex}{PurpleAir sensor’s index.}

\item{apiReadKey}{PurpleAir API read key with access to historical data.}

\item{startTimeStamp}{The beginning date in the format "YYYY-MM-DD HH:mm:ss".}

\item{endTimeStamp}{The end date in the format "YYYY-MM-DD" HH:mm:ss.}

\item{average}{The desired average in minutes, one of the following: "0" (real-time), "10", "30", "60", "360" (6 hour), "1440" (1 day).}

\item{fields}{The "Fields" parameter specifies which 'sensor data fields' to include in the response. See: https://community.purpleair.com/t/api-history-fields-descriptions/}
}
\value{
Dataframe of PurpleAir history data of a single sensor or multiple sensors.
}
\description{
R function to download historical data from PurpleAir sensors in the newer API.
Adapted from the getPurpleairApiHistory function
Source: https://github.com/willianflores/getPurpleairApiHistory
}
